% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dotplot_no_errors}
\alias{dotplot_no_errors}
\title{dotplot_no_errors}
\usage{
dotplot_no_errors(x, myTheme = simpleTheme(pch = 19, col = 1),
  qlabel = "Weighted average", add.text.to.qlabel = "",
  reorder.groups = TRUE, reordering = "decreasing", label.define = list(),
  reference.line = NULL, bar.color = 1, horizontal = TRUE, ...)
}
\arguments{
\item{x}{dataframe}

\item{myTheme}{default values for pch and col (todo)}

\item{qlabel}{text for axis label}

\item{add.text.to.qlabel}{additional text for axis label}

\item{reorder.groups}{default is TRUE}

\item{reordering}{order of sorting when reorder.groups is TRUE}

\item{label.define}{(todo)}

\item{reference.line}{value for the reference line. default is null}

\item{bar.color}{default is 1}

\item{horizontal}{layout of the dotplot. default is TRUE}
}
\description{
build a vertical dotplot without error bars. Used for
weighted averages
}
\author{
Rollie Parrish <rollie.parrish@ampa.org>
}

