% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gamutplot}
\alias{gamutplot}
\title{gamutplot}
\usage{
gamutplot(title = "test", mydata, reordering)
}
\arguments{
\item{title}{the main title for the plot}

\item{mydata}{data frame with group, numerator, denominator}

\item{reordering}{sort order for the table and plot}
}
\value{
a list - table has the raw data and plot has the dotplot
}
\description{
returns the GAMUT dotplot and table
}
\author{
Rollie Parrish <rollie.parrish@ampa.org>
}

