% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_CI}
\alias{get_CI}
\title{get_CI}
\usage{
get_CI(program_name, num, den)
}
\arguments{
\item{program_name}{Vector of names of the transport program}

\item{num}{Vector of numerators}

\item{den}{vector of denominators}
}
\description{
Determine the confidence intervals for each program
based on numerators/denominators
}
\author{
Rollie Parrish <rollie.parrish@ampa.org>
}

