% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_metrics_wm}
\alias{plot_metrics_wm}
\title{plot_metrics_wm}
\usage{
plot_metrics_wm(metric = "metric name", program_name, wavg, rate_overall,
  reordering = "decreasing", qlabel = "")
}
\arguments{
\item{metric}{Name of the GAMUT metric}

\item{program_name}{Vector of names of the transport program}

\item{wavg}{Weighted average}

\item{rate_overall}{the benchmark value from all records}

\item{reordering}{Sets the sorting order. Default is 'decreasing'}

\item{qlabel}{Default is 'Estimated rate'}
}
\description{
Returns a dot plot for a given metric. Based on weighted averages
}
\author{
Rollie Parrish <rollie.parrish@ampa.org>
}

