% Generated by roxygen2 (4.0.1): do not edit by hand
\name{save_metricData}
\alias{save_metricData}
\title{save_metricData}
\usage{
save_metricData()
}
\description{
get GAMUT data from REDCap, transform/aggregate by
individual program. Save as metricData to
GAMUT.Rdata file in the data folder
}
\author{
Rollie Parrish <rollie.parrish@ampa.org>
}

